Move all project-home-dir-setup code to config.py
